+++ gym_bb/tasks/monopod_base.py
@@ -78,7 +78,7 @@ class MonopodBase(task.Task, abc.ABC):
         # Create the action space
         action_space = gym.spaces.Box(low=np.array([-self.max_torque_upper_leg, -self.max_torque_lower_leg]),
                                       high=np.array([self.max_torque_upper_leg,  self.max_torque_lower_leg]),
-                                      dtype=np.float32)
+                                      dtype=np.float64)
         # Configure reset limits
         high = np.array([
             self._u_limit,
@@ -108,14 +108,14 @@ class MonopodBase(task.Task, abc.ABC):
         # Configure the reset space - this is used to check if it exists inside the reset space when deciding whether to reset.
         self.reset_space = gym.spaces.Box(low=low,
                                           high=high,
-                                          dtype=np.float32)
+                                          dtype=np.float64)
 
         # Configure the observation space
         obs_high = high.copy() * 1.2
         obs_low = low.copy() * 1.2
         observation_space = gym.spaces.Box(low=obs_low,
                                            high=obs_high,
-                                           dtype=np.float32)
+                                           dtype=np.float64)
 
         return action_space, observation_space
